Address Ian's comments, and improve StyleCop compliance.
