+++ Solutions/Ais.Net.Models.Specs/Ais.Net.Models.Specs.csproj
@@ -45,7 +45,4 @@
   <ItemGroup>
     <Folder Include="Steps\" />
   </ItemGroup>
-  <ItemGroup>
-    <PackageReference Update="Roslynator.Analyzers" Version="4.1.1" />
-  </ItemGroup>
 </Project>
