+++ Solutions/Ais.Net.Receiver.Host.Console/Ais.Net.Receiver.Host.Console.csproj
@@ -1,5 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
+  <Import Project="$(EndjinProjectPropsPath)" Condition="$(EndjinProjectPropsPath) != ''" />
+
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <TargetFramework>net6.0</TargetFramework>
@@ -25,6 +27,10 @@
     <PackageReference Include="Microsoft.Extensions.Configuration.FileExtensions" Version="[6.0.*,)" />
     <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="[6.0.*,)" />
     <PackageReference Include="System.Threading.Tasks.Dataflow" Version="[6.0.*,)" />
+    <PackageReference Include="Endjin.RecommendedPractices.GitHub" Version="2.1.2">
+      <PrivateAssets>all</PrivateAssets>
+      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+    </PackageReference>
   </ItemGroup>
 
   <ItemGroup>
