+++ Solutions/Ais.Net.Receiver/Ais/Net/Receiver/Configuration/LoggerVerbosity.cs
@@ -7,11 +7,33 @@
 #nullable disable annotations
 namespace Ais.Net.Receiver.Configuration;
 
+/// <summary>
+/// Defines the verbosity of the console output.
+/// </summary>
 public enum LoggerVerbosity
 {
+    /// <summary>
+    /// Essential only.
+    /// </summary>
     Quiet = 0,
+
+    /// <summary>
+    /// Statistics only.
+    /// </summary>
     Minimal = 1,
+
+    /// <summary>
+    /// Vessel Names and Positions.
+    /// </summary>
     Normal = 2,
+
+    /// <summary>
+    /// NMEA Sentences.
+    /// </summary>
     Detailed = 3,
+
+    /// <summary>
+    /// Messages and Errors.
+    /// </summary>
     Diagnostic = 4
 }
